Feature/sw 13/134 generic repository v2

+++ FoodplannerServices/Image/FoodImageService.cs
@@ -12,25 +12,34 @@ public class FoodImageService(IImageService imageService, IFoodImageRepository f
     public async Task<int> CreateFoodImage(int userid, Stream imageStream, string imageName, string imageType, long imageFileSize)
     {
         var imageId = await imageService.SaveImageAsync(userid, imageStream, imageType);
-        int foodImageId = await foodImageRepository.InsertImageAsync(
-            imageId.ToString(), 
-            userid, 
-            imageName, 
-            imageType, 
-            imageFileSize);
-        
+
+        FoodImage foodImage = new FoodImage
+        {
+            ImageId = imageId.ToString(),
+            UserId = userid,
+            ImageName = imageName,
+            ImageFileType = imageType,
+            Size = imageFileSize
+        };
+        int foodImageId = await foodImageRepository.InsertAsync(foodImage);
+
         return foodImageId;
     }
-    
+
     public async Task<FoodImageDTO> GetFoodImage(int foodImageId)
     {
-        var foodImage = await foodImageRepository.GetImageByIdAsync(foodImageId);
+        var foodImage = await foodImageRepository.GetByIdAsync(foodImageId);
         return _mapper.Map<FoodImageDTO>(foodImage);
     }
 
     public async Task<string> GetFoodImageLink(int foodImageId)
     {
-        var foodImage = await foodImageRepository.GetImageByIdAsync(foodImageId);
+
+        var foodImage = await foodImageRepository.GetByIdAsync(foodImageId);
+        if (foodImage == null)
+        {
+            throw new KeyNotFoundException($"FoodImage with id {foodImageId} not found");
+        }
         var foodImageLink = await imageService
             .LoadImagePresignedAsync(foodImage.UserId, Guid.Parse(foodImage.ImageId), foodImage.ImageFileType);
         if (foodImageLink == null)
@@ -42,8 +51,13 @@ public class FoodImageService(IImageService imageService, IFoodImageRepository f
 
     public async Task<bool> DeleteImage(int foodImageId)
     {
-        var foodImage = await foodImageRepository.GetImageByIdAsync(foodImageId);
-        await foodImageRepository.DeleteImageAsync(foodImageId);
+
+        var foodImage = await foodImageRepository.GetByIdAsync(foodImageId);
+        await foodImageRepository.DeleteAsync(foodImageId);
+        if (foodImage == null)
+        {
+            throw new KeyNotFoundException($"FoodImage with id {foodImageId} not found");
+        }
         return await imageService.DeleteImageAsync(foodImage.UserId, Guid.Parse(foodImage.ImageId), foodImage.ImageFileType);
     }
 }
